Stats: Move module-followers-page and module-followers
