+++ client/my-sites/stats/site.jsx
@@ -15,8 +15,7 @@ var StatsNavigation = require( './stats-navigation' ),
 	DatePicker = require( './stats-date-picker' ),
 	Comments = require( './stats-comments' ),
 	Countries = require( './stats-countries' ),
-	DatePicker = require( './stats-date-picker' ),
-	Followers = require( './module-followers' ),
+	Followers = require( './stats-followers' ),
 	ChartTabs = require( './stats-chart-tabs' ),
 	StatsModule = require( './stats-module' ),
 	statsStrings = require( './stats-strings' ),
